backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
cryptography==44.0.0
redis==5.2.1
confluent-kafka==2.6.1
elasticsearch==8.17.0
marshmallow==3.23.2
boto3==1.35.81
Pillow==11.1.0
Flask-Limiter==3.8.0
gunicorn==23.0.0
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
